Made chess its own cog and added time
